chore: pull more types into `types.rs`
